app/requirements.txt
numpy==1.26.4
streamlit==1.37.1
tensorflow==2.19.0
pillow